# testsnew and openapi backup/openapi_fixed.yaml
openapi: 3.0.3
info:
  title: Multi-Agent Orchestrator MCP Server - Cequence Deployment
  description: |
    Enterprise Multi-Agent Orchestrator MCP Server with Cequence AI Gateway Integration
    
    This deployment provides 16 total tools (11 standard + 5 advanced) with comprehensive 
    Cequence AI Gateway integration for enterprise-grade observability, security monitoring, 
    and performance analytics.
    
    🔒 **SECURITY IMPLEMENTATION:**
    - Secure Descope Access Key authentication with proper scope validation
    - Bearer token authentication with granular permissions
    - Vulnerability-free middleware with authentication bypass protections
    - Enterprise-grade security audit compliance
    
    **ADVANCED AGENT TOOLS (5 TOOLS):**
    1. Application Generator - Multi-agent orchestration for comprehensive application development
    2. Autonomous Architect - Dynamic system architecture generation with DAG optimization
    3. Proactive Quality Framework - Policy-driven quality assurance with predictive analysis
    4. Evolutionary Prompt Engine - Self-optimizing prompts using genetic algorithms
    5. Last Mile Cloud Agent - Multi-cloud deployment automation across AWS, Azure, GCP

    **STANDARD TOOLS (11 TOOLS):**
    1. Ping - Enhanced connectivity testing with scope validation
    2. Orchestrate Task - Multi-agent task coordination and execution
    3. Generate Architecture - AI-powered system design and planning
...

    **SECURITY & ANALYTICS FEATURES:**
...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /mcp/initialize:
    post:
      summary: "MCP Initialize"
      operationId: "mcp_initialize"
      description: |
        Enhanced MCP initialization with Cequence analytics capabilities
        
        Initializes MCP client-server connection with:
        - Advanced capability negotiation
        - Analytics integration setup
        - Performance monitoring initialization
        - Security context establishment
      tags: [MCP Protocol]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPInitializeRequest'
      responses:
        '200':
          description: MCP initialized with enhanced capabilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPInitializeResponse'
  /mcp/tools/list:
    post:
      summary: "MCP List Tools"
      operationId: "mcp_list_tools"
      description: |
        Comprehensive tool listing with analytics and capability insights
        
        Lists all available MCP tools including:
        - All 16 tools (5 advanced + 11 standard)
        - Tool performance metrics and analytics
        - Capability scoring and optimization insights
        - Usage patterns and recommendations
      tags: [MCP Protocol]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPListRequest'
      responses:
        '200':
          description: Complete tool list with analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPToolsListResponse'
  /mcp/tools/call:
    post:
      summary: "MCP Call Tool"
      operationId: "mcp_call_tool"
      description: |
        Universal tool execution with comprehensive analytics tracking
        
        Executes any MCP tool with:
        - Real-time performance monitoring
        - Security validation and compliance
        - Analytics data collection and insights
        - Optimization recommendations
      tags: [MCP Protocol]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPToolCallRequest'
      responses:
        '200':
          description: Tool executed with analytics insights
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPToolCallResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Descope Access Key authentication with Bearer tokens for enterprise MCP server access.
        
        This authentication method uses Descope Access Keys (JWT tokens) that integrate seamlessly 
        with Cequence AI Gateway in passthrough mode for comprehensive observability and security 
        monitoring while maintaining granular scope-based authorization.
        
        **Supported Scopes (for demo and documentation purposes):**
        
        **Advanced Agent Tools Scopes:**
        - `tools:advanced` - Access to all 5 advanced AI agent tools
        - `tools:autonomous` - Autonomous architect capabilities
        - `tools:evolutionary` - Evolutionary prompt optimization
        - `tools:proactive` - Proactive quality assurance
        - `tools:cloud` - Last mile cloud deployment
        
        **Standard Tools Scopes:**
        - `tools:ping` - Basic connectivity testing
        - `tools:generate` - Generation and orchestration tools
        - `tools:healing` - Autonomous code healing
        - `tools:infrastructure` - Infrastructure management tools
        - `tools:quality` - Quality assurance tools
        
        **Administrative Scopes:**
        - `admin:metrics` - System monitoring and analytics
        - `admin:analytics` - Advanced analytics access
        - `admin:config` - Configuration management
        - `admin:full` - Full administrative access
        
        **Integration with Cequence AI Gateway:**
        - Real-time analytics and performance monitoring
        - Security threat detection and prevention
        - Cost optimization and resource management
        - Compliance monitoring and automated reporting
        
        **Note:** Scope lists are for documentation/demo purposes; they are not declared as OpenAPI OAuth scopes.
...
  schemas:
...
    MCPInitializeRequest:
      type: object
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        id:
          type: string
        method:
          type: string
          example: "initialize"
        params:
          type: object
          properties:
            protocolVersion:
              type: string
              example: "2024-11-05"
            capabilities:
              type: object
            clientInfo:
              type: object
    MCPInitializeResponse:
      type: object
      properties:
        jsonrpc:
          type: string
        id:
          type: string
        result:
          type: object
          properties:
            protocolVersion:
              type: string
            capabilities:
              type: object
              properties:
                tools:
                  type: object
                  properties:
                    listChanged:
                      type: boolean
                    total_tools:
                      type: integer
                      example: 16
                    advanced_tools:
                      type: integer
                      example: 5
            serverInfo:
              type: object
              properties:
                name:
                  type: string
                  example: "multi-agent-orchestrator-cequence"
                version:
                  type: string
                  example: "3.0.0-CEQUENCE-ENHANCED"
    MCPListRequest:
      type: object
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        id:
          type: string
        method:
          type: string
          example: "tools/list"
    MCPToolsListResponse:
      type: object
      properties:
        jsonrpc:
          type: string
        id:
          type: string
        result:
          type: object
          properties:
            tools:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  description:
                    type: string
                  category:
                    type: string
                    enum: [advanced, standard, mcp_protocol]
                  inputSchema:
                    type: object
                  analytics_enabled:
                    type: boolean
                    example: true
    MCPToolCallRequest:
      type: object
      properties:
        jsonrpc:
          type: string
        id:
          type: string
        method:
          type: string
          example: "tools/call"
        params:
          type: object
          properties:
            name:
              type: string
            arguments:
              type: object
    MCPToolCallResponse:
      type: object
      properties:
        jsonrpc:
          type: string
        id:
          type: string
        result:
          type: object
          properties:
            content:
              type: array
              items:
                type: object
            cequence_execution_analytics:
              type: object
              properties:
                execution_time_ms:
                  type: number
                  format: float
                performance_score:
                  type: number
                  format: float
                resource_utilization:
                  type: object

    # === ADDITIONAL STANDARD TOOL SCHEMAS ===
...